+++ app/src/config/db.ts
@@ -1,0 +1,12 @@
+import * as mysql from "mysql2";
+
+const db: mysql.Connection = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PSWORD,
+  database: process.env.DB_DATABASE,
+});
+
+db.connect();
+
+export default db;
