+++ app/src/config/types.ts
@@ -1,0 +1,5 @@
+import * as core from "express-serve-static-core";
+
+export type query = core.Query;
+
+export type params = core.ParamsDictionary;
