--- app/src/models/utils/Error.js
@@ -1,11 +1,0 @@
-class Error {
-  static ctrl(msg, err) {
-    return {
-      isError: true,
-      errMsg: err,
-      clientMsg: msg,
-    };
-  }
-}
-
-module.exports = Error;
