+++ app/src/models/utils/Error.ts
@@ -1,0 +1,16 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+interface error {
+  isError: boolean;
+  errMsg: string;
+  clientMsg: string;
+}
+
+export default class {
+  static ctrl(msg: string, err: string): error {
+    return {
+      isError: true,
+      errMsg: err,
+      clientMsg: msg,
+    };
+  }
+}
