--- app/src/models/utils/String.js
@@ -1,7 +1,0 @@
-class String {
-  static makePrice(price) {
-    return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-  }
-}
-
-module.exports = String;
