+++ app/src/models/utils/String.ts
@@ -1,0 +1,7 @@
+class String {
+  static makePrice(price: number): string {
+    return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+  }
+}
+
+export default String;
